mv execution to functionality
